requirements.txt
# openai
# langchain
# langchain_pinecone
# huggingface
# ctransformers==0.2.5
# sentence-transformers==2.2.2
# pinecone
# langchain==0.0.225
# flask
# pypdf
# python-dotenv

# -e .
langchain==0.2.1
langchain-pinecone==0.1.1
langchain-community==0.2.1
langchain-core==0.2.1
pinecone-client==3.2.2
ctransformers==0.2.27
sentence-transformers==2.7.0
flask==3.0.3
pypdf==4.2.0
python-dotenv==1.0.1
-e .